% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.payment.R
\name{box.payment}
\alias{box.payment}
\title{box.payment}
\usage{
box.payment(df, P_type = "AMP")
}
\arguments{
\item{df}{Default dataset to use for the plot}

\item{P_type}{String indicating which column in the DRG data to plot. "AMP" plots
the Average Medicare Payment; "ATP" plots the Average Total Payments; "ACC" plots
the Average Covered Charges. The default value is "AMP".}
}
\value{
a box plot of the column user chose
}
\description{
This function produces a box plot of payment by DRG code.
The user can customize which payment type to draw in the plot
}
\examples{
box.payment(DRG_data, "ATP")


}
