% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRG.statistics.R
\name{DRG.statistics}
\alias{DRG.statistics}
\title{DRG.statistics}
\usage{
DRG.statistics(df, func = mean)
}
\arguments{
\item{df}{Default dataset to use for statistical analysis}

\item{func}{a function for the type of statistical value to return. "mean" returns
the mean value; "median" returns the median; "sd" returns the standard deviation
the default is "mean"}
}
\value{
A value of mean, median or standard deviation based on the user input
}
\description{
This function produces a statistic value feedback
}
\examples{
DRG.statistics(DRG_data, mean)

}
